Web/My First App/src/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
